# api/step-ci/schemas/openapi.yaml
openapi: 3.0.3
info:
  title: Todoist AI Agent API
  description: |
    AI-powered task management API following Google API Design Guide.
    
    ## Authentication
    Currently using environment-based API keys for Todoist and Claude.
    
    ## Resource Naming
    - Resources use kebab-case for URIs
    - Collections are plural (tasks, projects, labels)
    - Custom methods use :verb suffix pattern
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-app.vercel.app/api
    description: Production server (Vercel)
...
paths:
...
  /v1/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: List tasks
      description: Returns a list of tasks with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: priority
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 4
          description: Filter by priority (1-4)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          description: Maximum number of tasks to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset for pagination
        - name: completed
          in: query
          schema:
            type: boolean
            default: false
          description: Include completed tasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTasksResponse'
    post:
      tags: [tasks]
      operationId: createTask
      summary: Create a task
      description: Creates a new task in Todoist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
        description: Task ID
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get a task
      description: Returns a single task by ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Update a task
      description: Updates task properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Delete a task
      description: Deletes a task from storage
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===========================================================================
  # Tasks - Custom Methods
  # ===========================================================================
  /v1/tasks/sync:
    post:
      tags: [tasks]
      operationId: syncTasks
      summary: Sync with Todoist
      description: Synchronizes tasks with Todoist API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullSync:
                  type: boolean
                  default: false
                  description: Force full resync (ignore sync token)
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '400':
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/tasks/{taskId}/complete:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [tasks]
      operationId: completeTask
      summary: Complete a task
      description: Marks a task as complete in Todoist and local storage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actualMinutes:
                  type: integer
                  minimum: 1
                  description: Actual time spent in minutes
                context:
                  type: string
                  description: Context note for completion
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTaskResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/tasks/search:
    post:
      tags: [tasks]
      operationId: searchTasks
      summary: Search tasks
      description: Search tasks by content, description, or category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: Search query
                limit:
                  type: integer
                  default: 20
                  maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTasksResponse'
  /v1/tasks/batchUpdate:
    post:
      tags: [tasks]
      operationId: batchUpdateTasks
      summary: Batch update tasks
      description: Update multiple tasks in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [updates]
              properties:
                updates:
                  type: array
                  items:
                    type: object
                    required: [taskId, updates]
                    properties:
                      taskId:
                        type: string
                      updates:
                        $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Batch update results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchUpdateResponse'
  /v1/tasks/resetTestState:
    post:
      tags: [tasks]
      operationId: resetTestState
      summary: Reset test state
      description: Reset state for step-ci idempotency (test environment only)
      responses:
        '200':
          description: State reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Not allowed in production
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===========================================================================
  # Projects
  # ===========================================================================
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, content, projectId, priority, createdAt, updatedAt, isCompleted]
      properties:
        id:
          type: string
        content:
          type: string
        description:
          type: string
        projectId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 4
        labels:
          type: array
          items:
            type: string
        due:
          $ref: '#/components/schemas/TaskDue'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
        metadata:
          $ref: '#/components/schemas/TaskMetadata'
    TaskDue:
      type: object
      properties:
        date:
          type: string
          format: date
        datetime:
          type: string
          format: date-time
        string:
          type: string
        isRecurring:
          type: boolean
    TaskMetadata:
      type: object
      properties:
        category:
          type: string
        timeEstimate:
          type: string
        timeEstimateMinutes:
          type: integer
        size:
          type: string
          enum: [XS, S, M, L, XL]
        aiConfidence:
          type: number
          minimum: 0
          maximum: 1
    ListTasksResponse:
      type: object
      required: [tasks, totalCount]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        totalCount:
          type: integer
        nextPageToken:
          type: string
    CreateTaskRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        description:
          type: string
        projectId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 4
        dueString:
          type: string
        dueDate:
          type: string
          format: date
        labels:
          type: array
          items:
            type: string
    UpdateTaskRequest:
      type: object
      properties:
        content:
          type: string
        description:
          type: string
        projectId:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 4
        category:
          type: string
        labels:
          type: array
          items:
            type: string
    SyncResponse:
      type: object
      required: [synced, tasks, projects, labels, newTasks, duration]
      properties:
        synced:
          type: integer
        tasks:
          type: integer
        projects:
          type: integer
        labels:
          type: integer
        newTasks:
          type: integer
        duration:
          type: integer
          description: Duration in milliseconds
    CompleteTaskResponse:
      type: object
      required: [taskId, completedAt]
      properties:
        taskId:
          type: string
        completedAt:
          type: string
          format: date-time
        actualMinutes:
          type: integer
    SearchTasksResponse:
      type: object
      required: [results, query, totalMatches]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        query:
          type: string
        totalMatches:
          type: integer
    BatchUpdateResponse:
      type: object
      required: [updated, failed, results]
      properties:
        updated:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              taskId:
                type: string
              status:
                type: string
                enum: [success, failed]
              error:
                type: string

    # =========================================================================
    # Project Schemas
    # =========================================================================
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            code:
              type: integer
            message:
              type: string
            status:
              type: string
            details:
              type: array
              items:
                type: object